+++ packages/create/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @sap-ux/create
 
+## 0.5.2
+
+### Patch Changes
+
+-   Updated dependencies [aa2ff95b]
+    -   @sap-ux/adp-tooling@0.5.2
+    -   @sap-ux/app-config-writer@0.3.2
+
+## 0.5.1
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.5.1
+-   @sap-ux/app-config-writer@0.3.1
+
+## 0.5.0
+
+### Minor Changes
+
+-   1aa0fc43: Drop NodeJS 16 support, current supported versions NodeJS 18 and 20.
+
+### Patch Changes
+
+-   Updated dependencies [1aa0fc43]
+    -   @sap-ux/mockserver-config-writer@0.2.0
+    -   @sap-ux/app-config-writer@0.3.0
+    -   @sap-ux/cap-config-writer@0.2.0
+    -   @sap-ux/project-access@1.14.0
+    -   @sap-ux/adp-tooling@0.5.0
+    -   @sap-ux/logger@0.4.0
+
 ## 0.4.37
 
 ### Patch Changes